requirements.txt
setuptools==69.5.1
pytest==8.0.0
assertpy==1.1
pytest-mock==3.12.0
pytest-xdist==3.5.0
pytest-bdd==7.1.2
Appium-Python-Client==3.2.1
pywin32==306
PyAutoGUI==0.9.54
pillow==10.3.0
selenium==4.18.1
loguru==0.7.2
attrs==23.2.0
chromedriver-py==121.0.6167.85
geckodriver==0.0.1
psutil==5.9.8
allure-pytest==2.13.2
requests==2.31.0
lxml==5.1.0
jsonpath-ng==1.6.1
Cerberus==1.3.5